+++ packages/model-bank/package.json
@@ -29,7 +29,7 @@
     "./github": "./src/aiModels/github.ts",
     "./google": "./src/aiModels/google.ts",
     "./groq": "./src/aiModels/groq.ts",
-    "./lobehub": "./src/aiModels/lobehub.ts",
+    "./lobehub": "./src/aiModels/lobehub/index.ts",
     "./higress": "./src/aiModels/higress.ts",
     "./huggingface": "./src/aiModels/huggingface.ts",
     "./hunyuan": "./src/aiModels/hunyuan.ts",
