+++ packages/model-bank/src/aiModels/index.ts
@@ -31,7 +31,7 @@ import { default as infiniai } from './infiniai';
 import { default as internlm } from './internlm';
 import { default as jina } from './jina';
 import { default as lmstudio } from './lmstudio';
-import { default as lobehub } from './lobehub';
+import { default as lobehub } from './lobehub/index';
 import { default as minimax } from './minimax';
 import { default as mistral } from './mistral';
 import { default as modelscope } from './modelscope';
@@ -194,7 +194,7 @@ export { default as infiniai } from './infiniai';
 export { default as internlm } from './internlm';
 export { default as jina } from './jina';
 export { default as lmstudio } from './lmstudio';
-export { default as lobehub } from './lobehub';
+export { default as lobehub } from './lobehub/index';
 export { default as minimax } from './minimax';
 export { default as mistral } from './mistral';
 export { default as modelscope } from './modelscope';
