+++ packages/model-bank/src/aiModels/lobehub/chat/minimax.ts
@@ -1,0 +1,41 @@
+import { AIChatModelCard } from '../../../types/aiModel';
+
+export const minimaxChatModels: AIChatModelCard[] = [
+  {
+    abilities: {
+      functionCall: true,
+      reasoning: true,
+      search: true,
+    },
+    contextWindowTokens: 204_800,
+    description: 'Built for efficient coding and agent workflows.',
+    displayName: 'MiniMax M2',
+    enabled: true,
+    id: 'MiniMax-M2-Stable',
+    maxOutput: 131_072,
+    pricing: {
+      units: [
+        { name: 'textInput', rate: 0.3, strategy: 'fixed', unit: 'millionTokens' },
+        {
+          name: 'textInput_cacheWrite',
+          rate: 0.375,
+          strategy: 'fixed',
+          unit: 'millionTokens',
+        },
+        {
+          name: 'textInput_cacheRead',
+          rate: 0.03,
+          strategy: 'fixed',
+          unit: 'millionTokens',
+        },
+
+        { name: 'textOutput', rate: 1.2, strategy: 'fixed', unit: 'millionTokens' },
+      ],
+    },
+    releasedAt: '2025-10-27',
+    settings: {
+      searchImpl: 'params',
+    },
+    type: 'chat',
+  },
+];
