+++ packages/model-bank/src/aiModels/lobehub/chat/xai.ts
@@ -1,0 +1,32 @@
+import { AIChatModelCard } from '../../../types/aiModel';
+
+export const xaiChatModels: AIChatModelCard[] = [
+  {
+    abilities: {
+      functionCall: true,
+      reasoning: true,
+      search: true,
+      vision: true,
+    },
+    contextWindowTokens: 256_000,
+    description:
+      'Our newest and strongest flagship model, excelling in NLP, math, and reasoning—an ideal all-rounder.',
+    displayName: 'Grok 4',
+    enabled: true,
+    id: 'grok-4',
+    pricing: {
+      units: [
+        { name: 'textInput_cacheRead', rate: 0.75, strategy: 'fixed', unit: 'millionTokens' },
+        { name: 'textInput', rate: 3, strategy: 'fixed', unit: 'millionTokens' },
+        { name: 'textOutput', rate: 15, strategy: 'fixed', unit: 'millionTokens' },
+      ],
+    },
+    releasedAt: '2025-07-09',
+    settings: {
+      // reasoning_effort is not supported by grok-4. Specifying reasoning_effort parameter will get an error response.
+      // extendParams: ['reasoningEffort'],
+      searchImpl: 'params',
+    },
+    type: 'chat',
+  },
+];
