+++ packages/model-bank/src/aiModels/lobehub/index.ts
@@ -1,0 +1,10 @@
+import { lobehubChatModels } from './chat';
+import { lobehubImageModels } from './image';
+
+export { lobehubChatModels } from './chat';
+export { lobehubImageModels } from './image';
+export * from './utils';
+
+export const allModels = [...lobehubChatModels, ...lobehubImageModels];
+
+export default allModels;
